requirements.txt
Cython==0.29.13
pyliblo >= 0.9.1
termcolor==1.1.0
-e git+https://github.com/styk-tv/python-randomnames.git@beaa1fad993bf03ac5bc6f3ace2eaed119585f80#egg=randomnames
yamlordereddictloader==0.4.0
prettytable==0.7.2
mido==1.2.9
pretty-midi==0.2.8
ffmpeg-python==0.2.0
